Applies BrushOptionsPreferences so the brush options are recorded between sessions

+++ zen-android/src/main/java/zenproject/meditation/android/views/dialogs/brush/SizeChangedListener.java
@@ -1,5 +1,5 @@
 package zenproject.meditation.android.views.dialogs.brush;
 
 public interface SizeChangedListener {
-    void onSizeChanged(float pertentage);
+    void onSizeChanged(int pertentage);
 }
